Explicitly convert dones to tf.float32 data type to prevent error
This resolves #94

+++ tf2rl/algos/bi_res_ddpg.py
@@ -56,7 +56,7 @@ class BiResDDPG(DDPG):
     @tf.function
     def _compute_td_error_body(self, states, actions, next_states, rewards, dones):
         with tf.device(self.device):
-            not_dones = 1. - dones
+            not_dones = 1. - tf.cast(dones, dtype=tf.float32)
             # Compute standard TD error
             target_Q = self.critic_target(
                 [next_states, self.actor_target(next_states)])
